requirements.txt
aiogram
yt-dlp
ytmusicapi
python-dotenv
